changed get price to double

+++ src/main/java/com/pluralsight/Plaza/Pizzeria/toppings/PremiumTopping.java
@@ -17,7 +17,7 @@ public abstract class PremiumTopping extends Topping{
     }
 
     @Override
-    public int getPrice(int size) {
+    public double getPrice(int size) {
         return 0;
     }
 }
