+++ src/main/java/com/pluralsight/Plaza/Pizzeria/toppings/RegularTopping.java
@@ -7,7 +7,7 @@ public class RegularTopping extends Topping {
     }
 
     @Override
-    public int getPrice(int size) {
+    public double getPrice(int size) {
         return 0;
     }
 }
