+++ src/main/java/com/pluralsight/Plaza/Pizzeria/toppings/SouceTopping.java
@@ -7,7 +7,7 @@ public class SouceTopping extends Topping{
     }
 
     @Override
-    public int getPrice(int size) {
+    public double getPrice(int size) {
         return 0;
     }
 }
